+++ commands/download/mega.js
@@ -2,14 +2,14 @@
 const config = require('../../config');
 const { send } = require('../../helpers');
 
-const MESSAGE = '🚧 *MEGA downloder is currently still in development.*\n\nStay tuned for updates!';
+const MESSAGE = '🚧 *MEGA downloader is currently still in development.*\n\nStay tuned for updates!';
 
 module.exports = {
     name: 'mega',
     aliases: ['megadl', 'megadownload'],
     category: 'download',
     description: 'Feature in development',
-    usage: `${config.prefix}ai`,
+    usage: `${config.prefix}mega`,
 
     async execute(sock, msg, args) {
         const jid = msg.key.remoteJid;