+++ commands/download/pinterest.js
@@ -2,14 +2,14 @@
 const config = require('../../config');
 const { send } = require('../../helpers');
 
-const MESSAGE = '🚧 *Pinterest downloder is currently still in development.*\n\nStay tuned for updates!';
+const MESSAGE = '🚧 *Pinterest downloader is currently still in development.*\n\nStay tuned for updates!';
 
 module.exports = {
     name: 'pinterest',
     aliases: ['pin', 'pint'],
     category: 'download',
     description: 'Feature in development',
-    usage: `${config.prefix}ai`,
+    usage: `${config.prefix}pinterest`,
 
     async execute(sock, msg, args) {
         const jid = msg.key.remoteJid;